feat/javaapi: Refacotred code #43
