+++ src/main/resources/application.properties
@@ -1,19 +1,15 @@
 server.port=3000
 
-dataSource.database=ORACLE
-dataSource.driverClassName=oracle.jdbc.OracleDriver
-
-spring.oradatasource.driver-class-name=oracle.jdbc.OracleDriver
+oracle.datasource.driver-class-name=oracle.jdbc.OracleDriver
 # spring.datasource.url=jdbc:oracle:thin:@tcps://${ORACLEDB_HOST}:${ORACLEDB_PORT}/${ORACLEDB_SERVICENAME}
 # hardcode for now, as proxy account only works in dev
-spring.oradatasource.jdbc-url=jdbc:oracle:thin:@tcps://nrcdb03.bcgov:${ORACLEDB_PORT}/fortmp1.nrs.bcgov
-spring.oradatasource.username=${ORACLEDB_USER}
-spring.oradatasource.password=${ORACLEDB_PASSWORD}
+oracle.datasource.jdbc-url=jdbc:oracle:thin:@tcps://nrcdb03.bcgov:${ORACLEDB_PORT}/fortmp1.nrs.bcgov
+oracle.datasource.username=${ORACLEDB_USER}
+oracle.datasource.password=${ORACLEDB_PASSWORD}
 
-spring.pgdatasource.driver-class-name=org.postgresql.Driver
-spring.pgdatasource.jdbc-url=jdbc:postgresql://${POSTGRESQL_HOST}:5432/${POSTGRESQL_DATABASE}
-spring.pgdatasource.username=${POSTGRESQL_USER}
-spring.pgdatasource.password=${POSTGRESQL_PASSWORD}
+postgres.datasource.driver-class-name=org.postgresql.Driver
+postgres.datasource.jdbc-url=jdbc:postgresql://${POSTGRESQL_HOST}:5432/${POSTGRESQL_DATABASE}
+postgres.datasource.username=${POSTGRESQL_USER}
+postgres.datasource.password=${POSTGRESQL_PASSWORD}
 
 spring.jpa.show-sql=true
-# spring.jpa.database-platform=org.hibernate.dialect.Oracle12cDialect
